requirements.txt
# Django и его зависимости
Django==5.0.10
asgiref==3.7.2
sqlparse==0.4.4
tzdata==2023.3

# FastAPI и его зависимости
fastapi==0.115.6
uvicorn==0.32.1
pydantic==2.10.3
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.19

# База данных
psycopg2-binary==2.9.10
dj-database-url==2.3.0

# Облачное хранение
cloudinary==1.41.0

# Переменные окружения и конфигурация
python-dotenv==1.0.0
python-decouple==3.8

# HTTP и безопасность
requests==2.31.0
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
urllib3==2.1.0

# Веб-сервер и статические файлы
gunicorn
whitenoise==6.8.2

# Изображения и документы
Pillow==11.0.0
docutils==0.20.1

# Дополнительные зависимости
typing_extensions==4.12.2
click==8.1.7
colorama==0.4.6
MarkupSafe==3.0.2
Jinja2==3.1.4
starlette==0.41.3
anyio==4.7.0
sniffio==1.3.1
